# sentinel-ai/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SentinelAI API
  version: 0.1.0
  description: |
    Implemented: the security-engine scan contract (`/v1/scan`, internal) and the gateway routes for security scan/check,
    AI chat/generate, file scanning (`/v1/files/scan`), events, usage and policies (API-key auth), login/JWT, user and
    team management (`/v1/users`, `/v1/invitations`, `/v1/teams`) and per-organization provider credentials
    (`/v1/providers`). Routes marked (planned) are specified ahead of implementation.

    Management routes (users, teams, invitations, providers, API keys) require a USER session (JWT): an API key is never
    accepted for them, so a leaked key cannot create users, change roles or store provider credentials.
    Wire format is snake_case JSON. Responses never contain matched sensitive values.
servers:
  - url: http://localhost:4000
    description: API gateway
  - url: http://localhost:8001
    description: Security engine (implemented, internal only)
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /v1/security/scan:
    post:
      summary: Scan text under the organization policy (gateway; implemented)
      description: Requires permission `scan:use`. The scan is audited before the result is returned; if it cannot be audited the response is 503 and no text is returned.
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/GatewayScanRequest" } } } }
      responses:
        "200": { description: "decision + evidence + event_id (BLOCK also returns 200 with sanitized_text null)", content: { application/json: { schema: { $ref: "#/components/schemas/ScanResult" } } } }
        "401": { description: unauthorized }
        "403": { description: forbidden }
        "413": { description: payload too large }
        "422": { description: invalid request }
        "503": { description: audit unavailable (fail closed) }
  /v1/security/check:
    post:
      summary: Allow/deny check (gateway; implemented)
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/GatewayScanRequest" } } } }
      responses: { "200": { description: "{ allowed, decision, risk_level, failed_closed, event_id }" } }
  /v1/ai/chat:
    post:
      summary: Secure chat proxy - scan input, send only sanitized text, scan output (gateway; implemented for provider gemini)
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AiRequest" } } } }
      responses:
        "200": { description: model response after output policy, with security summary }
        "403": { description: blocked (policy, unknown provider, or fail-closed), content: { application/json: { schema: { $ref: "#/components/schemas/BlockedError" } } } }
        "502": { description: provider error (no content returned) }
        "503": { description: audit unavailable (fail closed) }
...
  /v1/ai/generate:
    post:
      summary: Secure generation proxy (gateway; implemented)
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [provider, prompt], properties: { provider: { type: string }, model: { type: string }, prompt: { type: string } } } } } }
      responses: { "200": { description: ok }, "403": { description: blocked }, "502": { description: provider error }, "503": { description: audit unavailable } }
...
  /v1/events/{id}:
    get: { summary: "Event detail (no raw content), including its explanation (implemented)", parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { "200": { description: ok } } }
  /v1/events/{id}/replay:
    post:
      summary: |
        Replay a recorded decision (implemented; events:read, live_judge needs evaluation:run). The caller supplies the
        original text (events never store it); the engine checks it against the recorded content hash and re-runs the
        decision with the recorded policy versions and judge verdict. The replay is audit-logged; nothing is stored.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { content: { application/json: { schema: { type: object, additionalProperties: false, required: [text], properties: { text: { type: string }, live_judge: { type: boolean } } } } } }
      responses:
        "200": { description: "content_matches, identical, recorded/replayed decision, differences, version diff" }
        "403": { description: live_judge without evaluation:run }
        "404": { description: not found }
        "409": { description: "event_has_no_explanation | policy_version_missing" }
        "502": { description: replay_failed }
  /v1/organization/ai-judge:
    get: { summary: "Organization switch for the external AI judge (implemented; policies:read)", responses: { "200": { description: "{ external_judge }" } } }
    put:
      summary: "Turn the external AI judge on or off for the organization (implemented; policies:write; audit-logged)"
      requestBody: { content: { application/json: { schema: { type: object, additionalProperties: false, required: [external_judge], properties: { external_judge: { type: boolean } } } } } }
      responses: { "200": { description: "{ external_judge }" } }
...
  /v1/policies:
    get: { summary: List policies (implemented), responses: { "200": { description: ok } } }
    post: { summary: Create policy version 1 (validated; unsafe ALLOW rules rejected) (implemented), requestBody: { content: { application/json: { schema: { $ref: "#/components/schemas/Policy" } } } }, responses: { "201": { description: created }, "422": { description: invalid } } }
  /v1/policies/{id}:
    put: { summary: Create a new active version of a policy (immutable history) (implemented), parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { "200": { description: ok } } }
    delete: { summary: Deactivate a policy (implemented), parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { "204": { description: deleted } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    apiKeyAuth: { type: apiKey, in: header, name: X-Sentinel-Api-Key }
    internalToken: { type: apiKey, in: header, name: X-Internal-Token }
  schemas:
    Severity: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
    Action: { type: string, enum: [ALLOW, HASH, MASK, TOKENIZE, REDACT, QUARANTINE, BLOCK] }
    Direction: { type: string, enum: [INPUT, OUTPUT] }
    RiskLevel: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
    EntityType:
      type: string
      enum: [NAME, LOCATION, EMAIL, PHONE, ADDRESS, DATE_OF_BIRTH, PAN, AADHAAR, PASSPORT, SSN, DRIVER_LICENSE, CREDIT_CARD, BANK_ACCOUNT, UPI, IFSC,
             API_KEY, AWS_CREDENTIAL, GOOGLE_CREDENTIAL, GITHUB_TOKEN, JWT, OAUTH_TOKEN, PASSWORD, PRIVATE_KEY, CONNECTION_STRING,
             HIGH_ENTROPY_SECRET, INTERNAL_URL, CONFIDENTIAL_MARKER, CUSTOM_CONFIDENTIAL,
             PROMPT_INJECTION, SYSTEM_PROMPT_EXTRACTION, JAILBREAK, DATA_EXFILTRATION]
...
    TimeWindow:
      type: object
      additionalProperties: false
      required: [start, end]
      properties: { start: { type: string, pattern: "^([01]\\d|2[0-3]):[0-5]\\d$" }, end: { type: string, pattern: "^([01]\\d|2[0-3]):[0-5]\\d$" } }
    RuleScope:
      type: object
      additionalProperties: false
      properties:
        users: { type: array, items: { type: string } }
        teams: { type: array, items: { type: string } }
        applications: { type: array, items: { type: string } }
        providers: { type: array, items: { type: string } }
        models: { type: array, items: { type: string } }
        environments: { type: array, items: { type: string } }
        ip_cidrs: { type: array, items: { type: string } }
        time_window_utc: { $ref: "#/components/schemas/TimeWindow" }
        directions: { type: array, items: { $ref: "#/components/schemas/Direction" } }
    PolicyRule:
      type: object
      additionalProperties: false
      required: [entity, action]
      description: ALLOW is rejected for credentials, payment cards and threat entities.
      properties:
        entity: { $ref: "#/components/schemas/EntityType" }
        action: { $ref: "#/components/schemas/Action" }
        severity: { $ref: "#/components/schemas/Severity" }
        min_confidence: { type: number, minimum: 0, maximum: 1, default: 0 }
        scope: { $ref: "#/components/schemas/RuleScope" }
    Policy:
      type: object
      additionalProperties: false
      required: [policy_id]
      properties:
        policy_id: { type: string, minLength: 1, maxLength: 128 }
        organization_id: { type: string, maxLength: 128 }
        version: { type: integer, minimum: 1, default: 1 }
        rules: { type: array, maxItems: 500, items: { $ref: "#/components/schemas/PolicyRule" } }
...
    Detection:
      type: object
      description: Evidence only. Never contains the matched value.
      required: [entity, confidence, severity, location, detector, detector_version]
      properties:
        entity: { $ref: "#/components/schemas/EntityType" }
        confidence: { type: number, minimum: 0, maximum: 1 }
        severity: { $ref: "#/components/schemas/Severity" }
        location: { type: object, required: [start, end], properties: { start: { type: integer }, end: { type: integer } } }
        detector: { type: string }
        detector_version: { type: string }
        value_digest: { type: [string, "null"], description: keyed HMAC prefix for correlation }
    RiskFactor:
      type: object
      required: [name, contribution, detail]
      properties: { name: { type: string }, contribution: { type: number }, detail: { type: string } }
    Risk:
      type: object
      required: [risk_score, risk_level, decision, factors]
      properties:
        risk_score: { type: integer, minimum: 0, maximum: 100 }
        risk_level: { $ref: "#/components/schemas/RiskLevel" }
        decision: { $ref: "#/components/schemas/Action" }
        factors: { type: array, items: { $ref: "#/components/schemas/RiskFactor" } }
    EntityAction:
      type: object
      required: [entity, action, count]
      properties: { entity: { $ref: "#/components/schemas/EntityType" }, action: { $ref: "#/components/schemas/Action" }, count: { type: integer } }
    ScanResult:
      type: object
      required: [request_id, decision, failed_closed, detections, entity_actions, risk, sanitized_text, policy_id, detector_version, latency_ms]
      properties:
        request_id: { type: string }
        decision: { $ref: "#/components/schemas/Action" }
        failed_closed: { type: boolean }
        fail_closed_reason: { type: [string, "null"] }
        detections: { type: array, items: { $ref: "#/components/schemas/Detection" } }
        entity_actions: { type: array, items: { $ref: "#/components/schemas/EntityAction" } }
        risk: { $ref: "#/components/schemas/Risk" }
        sanitized_text: { type: [string, "null"], description: null whenever the decision withholds content (BLOCK/QUARANTINE) }
        policy_id: { type: string }
        detector_version: { type: string }
        latency_ms: { type: number }
        explanation: { oneOf: [{ $ref: "#/components/schemas/Explanation" }, { type: "null" }] }
    Explanation:
      type: object
      description: |
        Why a decision was made. No content: detector names, scores, versions and a keyed HMAC of the scanned text. The
        judge's free-text `reason` is returned by scans but never stored (gateway strips it; a DB CHECK rejects it).
      required: [decided_by, tier, detectors_fired, policy, versions, content_hmac]
      properties:
        decided_by: { enum: [rules, classifier, judge, fail_closed] }
        tier: { type: integer, minimum: 1, maximum: 3 }
        detectors_fired: { type: array, items: { type: object, properties: { detector: { type: string }, entity: { type: string }, count: { type: integer }, max_confidence: { type: number }, tier: { type: integer } } } }
        classifier: { type: [object, "null"], properties: { model: { type: string }, score: { type: number }, threshold: { type: number }, band_low: { type: [number, "null"] }, band_high: { type: [number, "null"] }, band: { enum: [attack, uncertain, benign] }, windows_scored: { type: integer }, windows_total: { type: integer } } }
        judge: { type: [object, "null"], properties: { called: { type: boolean }, cached: { type: boolean }, skipped_reason: { type: [string, "null"] }, verdict: { enum: [attack, benign, null] }, category: { type: [string, "null"] }, confidence: { type: [number, "null"] }, reason: { type: [string, "null"] }, model: { type: [string, "null"] }, prompt_version: { type: [string, "null"] }, latency_ms: { type: [number, "null"] } } }
        policy: { type: object, properties: { policy_id: { type: string }, policy_version: { type: integer }, deciding_entity: { type: [string, "null"] }, deciding_action: { type: string }, source: { enum: [policy_rule, baseline, risk_escalation, no_detection, fail_closed] } } }
        versions: { type: object, additionalProperties: { type: string } }
        content_hmac: { type: string, pattern: "^[0-9a-f]{64}$" }
...
    GatewayScanRequest:
      type: object
      additionalProperties: false
      required: [text]
      properties:
        text: { type: string }
        direction: { $ref: "#/components/schemas/Direction" }
        context: { type: object, additionalProperties: false, properties: { application: { type: string }, team: { type: string }, environment: { type: string }, model: { type: string } } }
    AiRequest:
      type: object
      additionalProperties: false
      required: [provider, messages]
      properties:
        provider: { type: string, pattern: "^[a-z][a-z0-9_-]{1,31}$", examples: [gemini] }
        model: { type: string }
        messages:
          type: array
          minItems: 1
          items: { type: object, required: [role, content], properties: { role: { enum: [system, user, assistant] }, content: { type: string } } }
        application: { type: string }
        session_id: { type: string, pattern: "^[A-Za-z0-9_.:@-]{1,128}$", description: "Names a token-vault session so TOKENIZE tokens stay stable across turns and are hydrated in the reply. Namespaced per caller by the gateway." }
        hydrate: { type: boolean, description: "false returns tokens un-hydrated" }
...
    BlockedError:
      type: object
      required: [error, stage, decision, failed_closed, event_id]
      properties:
        error: { type: string, const: blocked }
        stage: { enum: [input, output] }
        decision: { $ref: "#/components/schemas/Action" }
        failed_closed: { type: boolean }
        reason: { type: [string, "null"] }
        event_id: { type: [string, "null"] }